watchllm/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "watchllm"
version = "0.1.0"
description = "WatchLLM monorepo (SDK and CLI)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"fastapi>=0.115.0,<1.0.0",
	"pydantic>=2.10.0,<3.0.0",
	"PyJWT[crypto]>=2.10.1,<3.0.0",
]

[project.scripts]
watchllm = "watchllm.cli.main:main"


